feat: extract functions and common styles to separate components
